requirements.txt
numpy
pandas
scipy
scikit-learn
torch
pyyaml
nbformat